// todo-webapp/src/generated/todo-api.ts
export interface components {
    schemas: {
        Todo: {
            id: string;
            text: string;
            done: boolean;
            /** @description username of the user who added it */
            createdBy: string;
            /** Format: date-time */
            createdAt: string;
        };
        TodoCreate: {
            text: string;
        };
        TodoUpdate: {
            done: boolean;
        };
        Error: {
            /** @description HTTP or application error code */
            code: number;
            /** @description short human-readable label */
            message: string;
            /** @description detailed explanation */
            description?: string;
            /** @description URI to documentation */
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}